Port to Gnome 45: initial efforts
This makes it work on Gnome 45, but it's not perfect as of now.

Imports cannot be wrapped around try-catch, rendering gi://NM a hard
dependency, at this point.

Certain values are made accessible to class instances via global
variables (those being `extension` in extenion.js; and `gettext_domain`,
`settings`, and `dir` in prefs.js). This works, as a proof of concept,
but needs to be implemented properly.

Legacy ByteArray has been replaced with TextDecoder.